Move PFS unit tests to correct directory
